Using UInt16 for ports in ConnectInfo

+++ Pantomime/sync/ConnectInfo.swift
@@ -17,9 +17,9 @@ class ConnectInfo {
     let imapAuthMethod: String
     let smtpAuthMethod: String
     let imapServerName: String
-    let imapServerPort: UInt32
+    let imapServerPort: UInt16
     let smtpServerName: String
-    let smtpServerPort: UInt32
+    let smtpServerPort: UInt16
     let imapTransport: ConnectionTransport
     let smtpTransport: ConnectionTransport
 
@@ -27,8 +27,8 @@ class ConnectInfo {
          imapUsername: String?, smtpUsername: String?,
          imapPassword: String, smtpPassword: String?,
          imapAuthMethod: String, smtpAuthMethod: String,
-         imapServerName: String, imapServerPort: UInt32, imapTransport: ConnectionTransport,
-         smtpServerName: String, smtpServerPort: UInt32, smtpTransport: ConnectionTransport) {
+         imapServerName: String, imapServerPort: UInt16, imapTransport: ConnectionTransport,
+         smtpServerName: String, smtpServerPort: UInt16, smtpTransport: ConnectionTransport) {
         self.email = email
         self.imapUsername = imapUsername
         self.smtpUsername = smtpUsername
@@ -47,9 +47,9 @@ class ConnectInfo {
     convenience init(email: String,
                      imapPassword: String,
                      imapAuthMethod: String, smtpAuthMethod: String,
-                     imapServerName: String, imapServerPort: UInt32,
+                     imapServerName: String, imapServerPort: UInt16,
                      imapTransport: ConnectionTransport,
-                     smtpServerName: String, smtpServerPort: UInt32,
+                     smtpServerName: String, smtpServerPort: UInt16,
                      smtpTransport: ConnectionTransport) {
         self.init(email: email, imapUsername: nil, smtpUsername: nil,
                   imapPassword: imapPassword, smtpPassword: nil,
